+++ ng/practices-ng-forms/src/lib/form-group-fluent-builder/form-group-types.ts
@@ -1,5 +1,14 @@
-import { FormGroupControlsValues, FormValueSignalsRecord } from '../utils/form.types';
+import { Signal } from '@angular/core';
+import { FormGroupControlsValues } from '../utils/form.types';
 import { FormControls } from './form-group-types.internal';
 
+export type FormValueSignalsRecordDeprecated<TFormValue> = {
+    /**
+     * @deprecated Use formGroup.valueSignal instead
+     */
+    [key in keyof TFormValue as key extends string ? `${key}Signal` : never]: Signal<TFormValue[key]>;
+};
+
 export type FormGroupValueWithSignals<TControls extends FormControls<TControls>> = FormGroupControlsValues<TControls> &
-    FormValueSignalsRecord<FormGroupControlsValues<TControls>>;
+    FormValueSignalsRecordDeprecated<FormGroupControlsValues<TControls>>;
+