+++ ng/practices-ng-forms/src/lib/utils/form.types.ts
@@ -109,3 +109,7 @@ export type FormValueSignalsRecord<TFormValue> = {
     [key in keyof TFormValue as key extends string ? `${key}Signal` : never]: Signal<TFormValue[key]>;
 };
 
+export type FormValueSignalsRecordWithoutPostfix<TFormValue> = {
+    [key in keyof TFormValue]: Signal<TFormValue[key]>;
+};
+